Update resolver consistency discussion
The language here was, in my view, more opinionated than necessary. I propose this change to make the tone more a factual discussion of the concern -- the reality is it's not absolutely necessary to have consistency, especially for a matter of seconds/minutes or even days in some cases, and, more importantly, it's not practical in many cases to literally deploy many subgraphs "simultaneously", so the language here is more about how to understand the issue, so you can avoid the problem and plan for any necessary migration to a new resolver behavior.